+++ pypesto/objective/__init__.py
@@ -4,12 +4,13 @@ Objective
 
 """
 
-from .objective import (ObjectiveOptions,
-                        Objective)
+from .objective import Objective
 from .amici_objective import AmiciObjective
-from .util import res_to_fval
+from .options import ObjectiveOptions
+from .util import res_to_chi2, sres_to_schi2
 
-__all__ = ["ObjectiveOptions",
-           "Objective",
-           "res_to_fval",
+__all__ = ["Objective",
+           "ObjectiveOptions",
+           "res_to_chi2",
+           "sres_to_schi2",
            "AmiciObjective"]
